remove cycles pass 1
