+++ front/lib/resources/data_source_view_resource.ts
@@ -435,3 +435,4 @@ export class DataSourceViewResource extends ResourceWithVault<DataSourceViewMode
     };
   }
 }
+
